+++ frontend/src/scss/import/_profile.scss
@@ -440,8 +440,17 @@
 /////
 /////
 /////
+
+.project_container {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+
+  gap: 50px;
+}
 .myProject {
-  width: 100%;
+  width: 85%;
+  margin: 0 auto;
 
   &_selectContainer {
     display: flex;
